Created Cubes App
A screen was created on which 3 pyramids are displayed. All elements were moved separately to widgets. The elements.dart file was created for the cubes, the pyramid.dart file was created for the pyramids themselves, and appBar was also moved to a separate file. All files (widgets) are connected to the cubes.dart file via part ''... The cubes.dart file itself is called via main.
dart.